separated all my classes to clean up the project directory and make it easier to get around
