+++ app/src/main/AndroidManifest.xml
@@ -13,40 +13,40 @@
         android:theme="@style/Theme.COMP2000_CW1"
         tools:targetApi="31">
         <activity
-            android:name=".admin_settings_page"
+            android:name=".Admin.admin_settings_page"
             android:exported="false" />
         <activity
-            android:name=".employee_request_holiday"
+            android:name=".Employee.employee_request_holiday"
             android:exported="true" />
         <activity
-            android:name=".settings_page"
+            android:name=".Shared.settings_page"
             android:exported="true" />
         <activity
-            android:name=".admin_view_employees"
+            android:name=".Admin.admin_view_employees"
             android:exported="true" />
         <activity
-            android:name=".admin_delete_employee"
+            android:name=".Admin.admin_delete_employee"
             android:exported="true" />
         <activity
-            android:name=".admin_holiday_requests"
+            android:name=".Admin.admin_holiday_requests"
             android:exported="true" />
         <activity
-            android:name=".admin_create_new_employee"
+            android:name=".Admin.admin_create_new_employee"
             android:exported="true" />
         <activity
-            android:name=".admin_main_hub"
+            android:name=".Admin.admin_main_hub"
             android:exported="true" />
         <activity
-            android:name=".employee_notifications"
+            android:name=".Employee.employee_notifications"
             android:exported="true" />
         <activity
-            android:name=".employee_edit_details"
+            android:name=".Employee.employee_edit_details"
             android:exported="true" />
         <activity
-            android:name=".employee_view_details"
+            android:name=".Employee.employee_view_details"
             android:exported="true" />
         <activity
-            android:name=".sign_in_page"
+            android:name=".Shared.sign_in_page"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
