--- app/src/main/java/EmployeeDetailModel.java
@@ -1,13 +1,0 @@
-public class EmployeeDetailModel {
-
-    //creating the getters and setters for all of the employee details
-    private int id;
-    private String firstname;
-    private String lastname;
-    private String email;
-    private String department;
-    private double salary;
-    private String joiningdate;
-    private int leaves;
-
-}
