+++ app/src/main/java/com/example/comp2000_cw1/Admin/admin_delete_employee.java
@@ -1,4 +1,4 @@
-package com.example.comp2000_cw1;
+package com.example.comp2000_cw1.Admin;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,27 +11,29 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-public class admin_settings_page extends AppCompatActivity {
+import com.example.comp2000_cw1.R;
+
+public class admin_delete_employee extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.admin_settings_page);
+        setContentView(R.layout.activity_admin_delete_employee);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
 
-        //code to find the button to navigate to the holidays page
-        Button returnHomeSet = findViewById(R.id.returnHomeSet);
 
-        //setting the onClickListener so that navigation to home holidays page is possible
-        returnHomeSet.setOnClickListener(new View.OnClickListener() {
+        Button deleteEmployee_returnBtn = findViewById(R.id.deleteEmployee_returnBtn);
+
+
+        deleteEmployee_returnBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(admin_settings_page.this, admin_main_hub.class);
+                Intent intent = new Intent(admin_delete_employee.this, admin_view_employees.class);
                 startActivity(intent);
             }
         });
