+++ app/src/main/java/com/example/comp2000_cw1/Admin/admin_holiday_requests.java
@@ -1,6 +1,9 @@
-package com.example.comp2000_cw1;
+package com.example.comp2000_cw1.Admin;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -8,22 +11,31 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-public class sign_in_page extends AppCompatActivity {
+import com.example.comp2000_cw1.R;
+
+public class admin_holiday_requests extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.sign_in_page);
+        setContentView(R.layout.activity_admin_holiday_requests);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
+
+
+        Button holidayRequest_returnBtn = findViewById(R.id.holidayRequest_returnBtn);
+
+
+        holidayRequest_returnBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(admin_holiday_requests.this, admin_main_hub.class);
+                startActivity(intent);
+            }
+        });
     }
-
-   
-
 }
-
-
