+++ app/src/main/java/com/example/comp2000_cw1/Employee/employee_notifications.java
@@ -1,4 +1,4 @@
-package com.example.comp2000_cw1;
+package com.example.comp2000_cw1.Employee;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,27 +11,29 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-public class admin_holiday_requests extends AppCompatActivity {
+import com.example.comp2000_cw1.R;
+
+public class employee_notifications extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_admin_holiday_requests);
+        setContentView(R.layout.employee_notifications);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
 
+        //code to find the button to navigate back to the home page
+        Button notifsHomeBtn = findViewById(R.id.notifsHomeBtn);
 
-        Button holidayRequest_returnBtn = findViewById(R.id.holidayRequest_returnBtn);
-
-
-        holidayRequest_returnBtn.setOnClickListener(new View.OnClickListener() {
+        //setting the onClickListener so that navigation to home page is possible
+        notifsHomeBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(admin_holiday_requests.this, admin_main_hub.class);
+                Intent intent = new Intent(employee_notifications.this, employee_view_details.class);
                 startActivity(intent);
             }
         });
