+++ app/src/main/java/com/example/comp2000_cw1/Employee/employee_view_details.java
@@ -1,4 +1,4 @@
-package com.example.comp2000_cw1;
+package com.example.comp2000_cw1.Employee;
 
 import android.os.Bundle;
 import android.content.Intent;
@@ -10,6 +10,9 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.example.comp2000_cw1.R;
+import com.example.comp2000_cw1.Shared.settings_page;
 
 public class employee_view_details extends AppCompatActivity {
