+++ app/src/main/java/com/example/comp2000_cw1/Shared/sign_in_page.java
@@ -1,9 +1,6 @@
-package com.example.comp2000_cw1;
+package com.example.comp2000_cw1.Shared;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -11,29 +8,24 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-public class admin_create_new_employee extends AppCompatActivity {
+import com.example.comp2000_cw1.R;
+
+public class sign_in_page extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_admin_create_new_employee);
+        setContentView(R.layout.sign_in_page);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
-
-
-        Button createEmployee_returnBtn = findViewById(R.id.createEmployee_returnBtn);
-
-
-        createEmployee_returnBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(admin_create_new_employee.this, admin_view_employees.class);
-                startActivity(intent);
-            }
-        });
     }
+
+   
+
 }
+
+
