+++ app/src/main/res/layout/activity_admin_create_new_employee.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".admin_create_new_employee">
+    tools:context=".Admin.admin_create_new_employee">
 
     <TextView
         android:id="@+id/createEmployeeTitle"
