+++ app/src/main/res/layout/activity_admin_delete_employee.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".admin_delete_employee">
+    tools:context=".Admin.admin_delete_employee">
 
     <TextView
         android:id="@+id/deleteEmployee_title"
