+++ app/src/main/res/layout/activity_admin_view_employees.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".admin_view_employees">
+    tools:context=".Admin.admin_view_employees">
 
     <TextView
         android:id="@+id/admin_viewEmployeesTitle"
