+++ app/src/main/res/layout/admin_settings_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".settings_page">
+    tools:context=".Shared.settings_page">
 
     <TextView
         android:id="@+id/settingsTitle"
