+++ app/src/main/res/layout/employee_edit_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".employee_view_details">
+    tools:context=".Employee.employee_view_details">
 
 
     <EditText
