+++ app/src/main/res/layout/employee_notifications.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".employee_notifications">
+    tools:context=".Employee.employee_notifications">
 
     <TextView
         android:id="@+id/notificationsTitle"
