+++ app/src/main/res/layout/employee_request_holiday.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".employee_request_holiday"
+    tools:context=".Employee.employee_request_holiday"
     tools:ignore="VisualLintBounds">
 
     <CalendarView
