+++ app/src/main/res/layout/sign_in_page.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".sign_in_page">
+    tools:context=".Shared.sign_in_page">
 
 
     <TextView
